fix: honor signed event-driven exits
Why: Cloud review found trailing cat signatures made finalRoutingSlot select the signature paragraph, so signed External Wait: event-driven exits were missed by remedial/verdict/void guards.

[砚砚/GPT-5.5🐾]
